requirements.txt
fastapi
uvicorn
pandas
google-generativeai
pydantic